use connect popup view
