+++ packages/next-common/components/login/content.jsx
@@ -1,29 +1,18 @@
-import { useEffect } from "react";
 import LoginWeb3LoginContent from "./web3LoginContent";
 import LoginAccountLoginContent from "./accountLoginContent";
 import LoginEVMContent from "./evm/content";
-import { useSelector } from "react-redux";
-import {
-  connectPopupViewSelector,
-  setConnectPopupView,
-} from "next-common/store/reducers/connectPopupSlice";
-import { useDispatch } from "react-redux";
-import {
-  CONNECT_POPUP_DEFAULT_VIEW,
-  CONNECT_POPUP_VIEWS,
-} from "next-common/utils/constants";
+import { CONNECT_POPUP_VIEWS } from "next-common/utils/constants";
+import { useConnectPopupView } from "next-common/hooks/connect/useConnectPopupView";
+import { useEffect } from "react";
 
 /**
  * @description used in login popup and login page
  */
 export default function LoginContent() {
-  const view = useSelector(connectPopupViewSelector);
-  const dispatch = useDispatch();
+  const [view, , reset] = useConnectPopupView();
 
   useEffect(() => {
-    return () => {
-      dispatch(setConnectPopupView(CONNECT_POPUP_DEFAULT_VIEW));
-    };
+    return reset;
   }, []);
 
   return (
